move descriptor to utils
