+++ src/cryptoadvance/specter/devices/bitcoin_core.py
@@ -1,9 +1,9 @@
 import os
 import shutil
 from mnemonic import Mnemonic
-from hwilib.descriptor import AddChecksum
 from ..device import Device
 from ..helpers import alias
+from ..util.descriptor import AddChecksum
 from ..util.base58 import encode_base58_checksum, decode_base58
 from ..util.xpub import get_xpub_fingerprint, convert_xpub_prefix
 from ..key import Key
