+++ src/cryptoadvance/specter/util/descriptor.py
